Modifications to the localization and person routes

+++ src/@core/application/use-cases/localization/delete-city.usecase.ts
@@ -1,11 +1,14 @@
 import { ILocalization } from 'src/@core/domain/contracts/localization-repository.interface';
 import UseCase from '../usecase';
+import { NotFoundError } from 'src/@core/shared/domain/errors/not-found.error';
 
 export namespace CityDelete {
   export class Usecase implements UseCase<Input, Output> {
     constructor(private repo: ILocalization) {}
 
     async execute(input: Input): Output {
+      input.name = input.name.toUpperCase();
+      if (!await this.repo.getCity(input.name)) throw new NotFoundError('City not found');
       return await this.repo.deleteCity(input.name);
     }
   }
@@ -16,6 +19,5 @@ export namespace CityDelete {
 
   export type Output = Promise<{
     name: string
-    deleted: boolean
   }>;
 }
